refactor: Separate theme utilities from config for non-programmer accessibility - Moved all utility functions from theme-config.js to theme-utils.js - Cleaned theme-config.js to only contain variable definitions - Enhanced documentation for non-technical contributors - Updated imports in theme-manager.js - Reduced theme-config.js complexity by 44% (437→243 lines) - All functionality preserved, tested in both v12/v13 containers
